Modernize type annotations and prepare test infrastructure
- Convert to modern type syntax (X | Y, collections.abc imports)
- Fix type compatibility issues between EmptySection and RawSection
- Add test scaffolding files with implementation TODOs
- Improve function signatures and parameter handling

+++ tests/test_benchmark.py
@@ -12,7 +12,7 @@ def test_benchmark():
 
     block = Block('iron_block')
 
-    def func(x: int, z: int) -> int:
+    def func(x: float, z: float) -> float:
         return math.sin(x * x + z * z) * 0.5 + 0.5
 
     w = 256
@@ -47,7 +47,7 @@ def test_raw_section():
     air = Block('air')
     palette = (air, block)
 
-    def func(x: int, z: int) -> int:
+    def func(x: float, z: float) -> float:
         return math.sin(x * x + z * z) * 0.5 + 0.5
 
     w = 256
